simplify client creation API
This changes the `NewMock` function to simplify the creation of the mock client
and to hide the details of the implementation from the user-facing API.  Instead
of taking a plugin connection as a parameter, the connection is established by
the `NewMock` function.  The plugin related options are exposed via options in a
new `mock` package.  The separate `NewMockFrom` function is removed and replaced
with an option (`WithSnapshotReader`) that has the same effect.  Finally, the
`NewRPC` function is also updated to take a slice of `Config` in place of the
variadic form to match `NewMock`.