Sam at luther/add private optimization (#16)
The idea here is to combine the encode transaction directly into the
call transaction, and instead pass the message that was going to be
encoded as transient data. Substrate would then immediately do the
encode/decode step (to ensure the DSR gets created) prior to evaluating
the wrapped call transaction.

One potential issue is that the transactions become atomic, where a DSR
would not be created if the wrapped call fails, but perhaps this is
desirable.